requirements.txt
ollama>=0.1.0
psycopg2-binary>=2.9.9
sentence-transformers>=2.2.0
numpy>=1.26.0
PyPDF2>=3.0.0
python-dotenv>=1.0.1
transformers>=4.30.0
torch>=2.0.0